Incapsulated implementation details of validator
